Cleaning up into additional files
